Fixed date parsing using .toJSON()
googlechart?data=true return date values looking like "U2015/0/12 16:20:0".
That's exactly what it is supposed to according to the code but that date,
even with a stipped "U" is not parsable by new Date().
Mostly, I guess, because of getMonth() returns 0-11 and not 1-12 and parsing
of month 0 is not doable.

Chrome gives
new Date("2015/0/12 16:20:0")
> Invalid Date

.toJSON() should be browser compatible according to
https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Date/toJSON